feat(release): implement worker-based audio metadata extraction and refactor related functions

+++ src/lib/helpers/metadata.ts
@@ -1,5 +1,5 @@
 import { parseBlob, selectCover, type IAudioMetadata, type IPicture } from 'music-metadata';
-import ParseAudioMetadata from '$lib/helpers/workers/parseAudioMetadata.ts?worker';
+import ParseAudioMetadata from '$lib/helpers/workers/ParseAudioMetadata.ts?worker';
 
 export type FileMetadata = Pick<IAudioMetadata, 'common'|'format'> & {
     title: string|null;
@@ -8,7 +8,7 @@ export type FileMetadata = Pick<IAudioMetadata, 'common'|'format'> & {
     [key: string]: unknown;
 };
 
-export async function extractFileMetadata(file: File, useWorker?: boolean): Promise<FileMetadata> {
+export async function extractFileMetadata(file: File, useWorker?: boolean|Worker): Promise<FileMetadata> {
     if (!useWorker) {
         const { common, format } = await parseBlob(file);
 
@@ -21,19 +21,22 @@ export async function extractFileMetadata(file: File, useWorker?: boolean): Prom
         };
     }
 
-    const worker = new ParseAudioMetadata();
-
-    console.log('Extracting metadata for file:', file.name);
-
+    const worker = useWorker instanceof Worker ? useWorker : new ParseAudioMetadata();
     const promise = new Promise<FileMetadata>((resolve, reject) => {
         worker.onmessage = event => {
             resolve(event.data);
-            worker.terminate();
+
+            if (useWorker == true) {
+                worker.terminate();
+            }
         };
 
         worker.onerror = error => {
             reject(error);
-            worker.terminate();
+
+            if (useWorker == true) {
+                worker.terminate();
+            }
         };
     });
